move viewmodel class
