+++ nav-app/src/app/classes/index.ts
@@ -7,3 +7,4 @@ export { LayoutOptions } from "./layout-options";
 export { Link } from "./link";
 export { Metadata } from "./metadata";
 export { TechniqueVM } from "./technique-vm";
+export { ViewModel } from "./view-model";
