+++ nav-app/src/app/matrix/cell.ts
@@ -1,7 +1,7 @@
 import { Input, Directive } from '@angular/core';
 import { DataService } from '../services/data.service';
 import { Technique, Tactic } from '../classes/stix';
-import { ViewModel } from '../services/viewmodels.service';
+import { ViewModel } from '../classes';
 import { getCookie, hasCookie } from "../utils/cookies";
 import { ConfigService } from '../services/config.service';
 import tinycolor from "tinycolor2";
