+++ nav-app/src/app/search-and-multiselect/search-and-multiselect.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit, Input, ViewEncapsulation } from '@angular/core';
-import { ViewModel, ViewModelsService } from '../services/viewmodels.service';
+import { ViewModelsService } from '../services/viewmodels.service';
 import { StixObject, Group, Mitigation, Software, Technique, Campaign } from '../classes/stix';
 import { DataService } from '../services/data.service';
+import { ViewModel } from '../classes';
 
 @Component({
     selector: 'app-search-and-multiselect',
